requirements1.txt
google-cloud-storage
bs4 >= 0.0.1
requests>=2.20.0
lxml >=4.8.0